// src/modules/cache/enums.ts
export enum CacheTable {
  DB_STATE = 'db/state',
  INDEXER_ACCOUNT = 'indexer/account',
  INDEXER_ACCOUNT_ASSETS = 'indexer/accountAssets',
  INDEXER_ACCOUNT_APPLICATIONS = 'indexer/accountApplications',
  INDEXER_ACCOUNT_TRANSACTIONS = 'indexer/accountTransactions',
  INDEXER_APPLICATION = 'indexer/application',
  INDEXER_APPLICATION_BOX = 'indexer/applicationBox',
  INDEXER_APPLICATION_BOXES = 'indexer/applicationBoxes',
  INDEXER_ASSET = 'indexer/asset',
  INDEXER_ASSET_BALANCES = 'indexer/assetBalances',
  INDEXER_ASSET_TRANSACTIONS = 'indexer/assetTransactions',
  INDEXER_BLOCK = 'indexer/block',
  INDEXER_TXN = 'indexer/txn',
  NODE_ACCOUNT = 'node/account',
  NODE_ACCOUNT_APPLICATION = 'node/accountApplication',
  NODE_ACCOUNT_ASSET = 'node/accountAsset',
  NODE_BLOCK = 'node/block',
  NODE_BLOCK_PROOF = 'node/blockProof',
  NODE_BLOCK_TRANSACTION_PROOF = 'node/blockTransactionProof',
  NODE_TEAL = 'node/teal',
  INDEXER_APPLICATIONS = 'indexer/applications',
  INDEXER_ACCOUNTS = 'indexer/accounts',
  INDEXER_ASSETS = 'indexer/assets',
  INDEXER_TXNS = 'indexer/txns',
  NFD_LOOKUP = 'nfd/lookup',
  NFD_SEARCH = 'nfd/search',
  MEDIAS_ASSET = 'medias/asset',
}